Recro/requirements.txt
#Working requirements.txt 


# Web Framework
fastapi==0.115.6
uvicorn==0.34.0
starlette==0.41.3
python-dotenv==1.2.1
python-multipart==0.0.20

# Auth
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# Azure
azure-cognitiveservices-speech==1.47.0
azure-core==1.38.0
azure-storage-blob==12.24.0

# LangChain ecosystem
langchain==0.4.0.dev0
langchain-core==0.4.0.dev0
langchain-community==0.3.31
langchain-openai==0.4.0.dev0
langchain-groq==0.3.8
langsmith==0.4.37
langgraph==0.6.11

# Tokenization
tiktoken==0.12.0

# LLM clients
openai==1.109.1
groq==0.37.1

# Audio / Speech
openai-whisper==20250625
pydub==0.25.1
soundfile==0.13.1
SpeechRecognition==3.14.5
sounddevice==0.5.5

# Data
pandas==2.3.3
numpy==1.26.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
orjson==3.11.5

# Streamlit UI
streamlit==1.50.0
streamlit-audiorecorder==0.0.6
streamlit-autorefresh==1.0.1
streamlit_mic_recorder==0.0.8
altair==5.5.0
pydeck==0.9.1

# ML / Transformers
transformers==4.57.6
sentence-transformers==5.1.2
torch==2.8.0
torchvision==0.23.0
scikit-learn==1.6.1

# Computer Vision
opencv-python==4.10.0.84
mediapipe==0.10.15
ultralytics==8.4.12

# Utilities
fpdf==1.7.2
pypdf==6.6.2
python-docx==1.2.0
svgwrite==1.4.3
matplotlib==3.9.4
polars==1.36.1
pyarrow==21.0.0
GitPython==3.1.46
psutil==7.2.2

pydantic[email]==2.12.5
faster-whisper==1.2.1
bcrypt<4.1